StepTalk framework: int->NSInteger transition
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/steptalk/trunk@36657 72102866-910b-0410-8b05-ffd578937521

+++ Frameworks/StepTalk/NSInvocation+additions.h
@@ -34,8 +34,8 @@ id   STObjectFromValueOfType(void *value, const char *type);
 + invocationWithTarget:(id)target selectorName:(NSString *)selectorName;
 + invocationWithTarget:(id)target selector:(SEL)selector;
 
-- (void)setArgumentAsObject:(id)anObject atIndex:(int)anIndex;
-- (id)getArgumentAsObjectAtIndex:(int)anIndex;
+- (void)setArgumentAsObject:(id)anObject atIndex:(NSInteger)anIndex;
+- (id)getArgumentAsObjectAtIndex:(NSInteger)anIndex;
 - (id)returnValueAsObject;
 @end
