+++ Frameworks/StepTalk/STActor.m
@@ -169,8 +169,8 @@ some other, more clever mechanism. */
     NSString       *methodName = NSStringFromSelector([invocation selector]);
     NSMutableArray *args;
     id              arg;
-    int             index;
-    int             count;
+    NSUInteger      index;
+    NSUInteger      count;
     id              retval = nil;
 
     method = [methodDictionary objectForKey:methodName];
